requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

certifi==2021.5.30

charset-normalizer==2.0.4; python_version >= '3'

idna==3.2; python_version >= '3'

pytz==2021.1

pyyaml==5.4.1

requests==2.26.0

todoist-python==8.1.3

urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'

